Added a one-dimensional solver (ie find x satisfying f(x) = 0) implementation Added bootstrap constructor with parameters and initialization functional in opposite order to match other distributions Added method to invert jackknife resampling (valid only if linear ops have been performed) Added computation of covariance for superMultiDistribution phaseShift calculation no longer assumes integer lattice size In fit_pipi_multiop_comoving_gparity_bootstrap, added ability to print a single sample of data for debug In analyze_ktopipi_gparity: 	Added calculation of ImA0 using optimal combination of data 	Added ability to compute step-scaling matrix including G1 Improved testing of superMultiDistribution

+++ include/physics/luscherzeta.h
@@ -265,7 +265,7 @@ double dispersionRelationGetEnergyFromEnergy(const DispersionRelation disp, GSLv
 
 //Energy E is in the *lab frame*
 //Phase shift is in degrees
-inline double phaseShiftZ(const LuscherZeta &zeta, double E, const double m, const int L, const DispersionRelation dispn = DispersionRelation::Continuum, const double zero_tol =1e-10){
+inline double phaseShiftZ(const LuscherZeta &zeta, double E, const double m, const double L, const DispersionRelation dispn = DispersionRelation::Continuum, const double zero_tol =1e-10){
   const GSLvector &d = zeta.getd();
 
   //Get CofM energy
@@ -288,7 +288,7 @@ inline double phaseShiftZ(const LuscherZeta &zeta, double E, const double m, con
 }
 
 //d= Pcm /(2pi/L) 
-inline double phaseShift(const double E, const double m, const int L, const std::array<int,3> &twists = {0,0,0}, const std::array<double,3> &d = {0.,0.,0.}, const DispersionRelation dispn = DispersionRelation::Continuum){
+inline double phaseShift(const double E, const double m, const double L, const std::array<int,3> &twists = {0,0,0}, const std::array<double,3> &d = {0.,0.,0.}, const DispersionRelation dispn = DispersionRelation::Continuum){
   LuscherZeta zeta(twists, d);
   return phaseShiftZ(zeta,E,m,L,dispn);
 }
